refactor: react components frontend structure
