+++ nextjs/src/app/game/TwoPlayerBoard.tsx
