+++ nextjs/src/app/layout.tsx
@@ -1,10 +1,10 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import NavigationBar from "../componets/NavigationBar";
+import NavigationBar from "../componets/sidebar/NavigationBar";
 import { getSession } from "@/lib/auth/session";
 import AuthProvider from "@/componets/AuthProvider";
-import SidebarActionsProvider from "@/componets/SidebarActionsProvider";
+import SidebarActionsProvider from "@/componets/sidebar/SidebarActionsProvider";
 import Footer from "./Footer";
 
 const geistSans = Geist({
